# openapi.yaml
openapi: 3.1.0
info:
  title: PabandiOS API
  version: 1.0.0
  description: The trust operating system for service businesses — bookings, freight, property, CRM, and finance with AI-powered reliability scoring and Solana escrow.
  contact:
    name: PabandiOS Agents
    email: agents@pabandi.com
servers:
  - url: https://api.pabandi.com
    description: Production
paths:
...
  /api/v1/booking/create:
    post:
      tags: [Booking]
      summary: Create a booking with escrow deposit
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [businessId, reservationDate, reservationTime, numberOfGuests]
              properties:
                businessId: { type: string }
                reservationDate: { type: string, format: date }
                reservationTime: { type: string }
                numberOfGuests: { type: integer }
                depositAmount: { type: number }
                specialRequests: { type: string }
                paymentMethod: { type: string }
                intentMandate: { type: object }
                cartMandate: { type: object }
                paymentMandate: { type: object }
      responses:
        '201':
          description: Booking created with deposit
        '401':
          description: AP2 mandates missing
        '402':
          description: Payment required (x402)
  /api/v1/booking/confirm:
    post:
      tags: [Booking]
      summary: Confirm booking payment and create escrow
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bookingReference: { type: string }
                paylioPaymentId: { type: string }
      responses:
        '200':
          description: Escrow created
  /api/v1/booking/{reference}/status:
    get:
      tags: [Booking]
      summary: Get booking status
      security:
        - bearerAuth: []
      parameters:
        - name: reference
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Booking status
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT